# perf: make parse_scheme slightly faster