Avoids a check that's already covered by another check and avoids doing a `is_ascii_uppercase()` check in `to_ascii_lowercase()` when we know the character is lowercase.

Before:

```
test short          ... bench:         189 ns/iter (+/- 6) = 132 MB/s
```

After:

```
test short          ... bench:         182 ns/iter (+/- 5) = 137 MB/s
```